# src/linkservice/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Linkservice. Service for manage short link.",
        "title": "URL shortener",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/links": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "link"
                ],
                "summary": "List links",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Link hashes",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.singleLinkResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create short url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "link"
                ],
                "summary": "Create short url",
                "parameters": [
                    {
                        "description": "Link URL",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.singleLinkResponse"
                        }
                    }
                }
            }
        },
        "/links/search": {
            "get": {
                "description": "get links by parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "link"
                ],
                "summary": "Search links",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of links in page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "url/hash/created_at",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc/desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search for url",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.searchLinkResponse"
                            }
                        }
                    }
                }
            }
        },
        "/links/{hash}": {
            "get": {
                "description": "get link info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "link"
                ],
                "summary": "Info about link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Link hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.infoLinkResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.infoLinkResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "redirects": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handler.searchLinkResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.singleLinkResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "handler.singleLinkResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}